first work on dotdot zcl loading
